backend/tracking-service/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9
python-dotenv>=1.0.1
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.0.0
pika>=1.3.0
python-multipart
requests
apscheduler
redis>=4.5.0
geopy>=2.3.0
